compiler/arm/{arith,parms}: Define static-funs for TWO-ARG-[<>].
  * The first two cases that we encounter, the observed symptom
being a stack overflow with all the stack frames being for the
same function (in these cases, either < or >).

  * A full call to < or > goes to the defined function in
code/numbers.  This function does a pairwise operation to break it
down to where the compiler can be somewhat intelligent.

  * A pairwise call to < or > with sufficient and compatible type
information and a defined VOP will be open-coded.  Otherwise, a
full call to < or > is compiled...  Unless there's a defined
static-fun, in which case a slightly different call is assembled
to the TWO-ARG-whatever version, which does a type-dispatch to
trigger the appropriate open-coded version (or other noise if the
types really aren't compatible, such as a FIXNUM against a BIGNUM).

  * Of course, if the VOP for an open-coded version is missing
(such as, say, > for two SINGLE-FLOAT args) then we end up with a
blown control stack showing all the stack frames for the same
TWO-ARG-whatever function.

  * In the latter case (blown stack, TWO-ARG-whatever), fire up
GDB, set a breakpoint at the start of the function in question,
and check on the args to see what they are.  Or just desk-check
the backend to see which (if any) relevant VOPs are missing.