compiler/{ir1opt,node}: Allow forcing a CAST node to always exist.
  * At the IR2 level, CAST nodes have the rather wonderful effect
of moving values from one LVAR to another.  This is occasionally a
useful thing to do in and of itself, with the drawback being that
a CAST which the compiler proves will never fail its typecheck
ends up getting eliminated during IR1OPT, outright replacing one
LVAR with the other.

  * Allow marking a CAST node as "never delete", meaning that even
if the compiler COULD prove that it will never fail its typecheck
the node must remain in place.  This is a bit of a KLUDGE, but too
useful of a thing not to do, and far less effort than adding a
whole new node type to get the same IR2 effect.