code/debug-int,runtime/arm-assem.S: Fix undefined-function backtrace.
  * Two things here.  First is that MAKE-LISP-OBJ didn't recognize
undefined_tramp as a valid object.  Easily enough fixed.  Second
is that the undefined_tramp wasn't initializing reg_CODE before
trapping, leaving the debugger looking in the wrong place for the
error trap location.  Also easily enough fixed.

  * Unlike PPC, we always have fully-initialized call frames
constructed on the caller side, so we don't need to do the rest of
the fixups that PPC does.