compiler/ltn: Robustify handling of IR2-blocks during LTN.
  * In ANNOTATE-UNKNOWN-VALUES-LVAR, it is assumed that any node
that is the DEST for an LVAR is in an IR1 block that already has
an associated IR2 block.  This may have been true historically,
but is no longer the case if it ever was.

  * Allocate IR2-BLOCK structures in an initial pre-pass over the
component, and ensure that new IR2-BLOCK structures are allocated
as needed whenever we split an IR1 block.

  * When we split an IR1 block, make sure that information about
popped LVARs is distributed correctly between the corresponding
IR2 blocks.

  * This fixes the known regression from the nonlinear LVAR fixes,
and should generally be a whole lot more robust.