no more kludge-nondeterministic-catch-block-size
It turns out that the only reason for delaying objdef so late in
the build was the need to inform the compiler's info database
about functions to access slots.  To resolve the ordering issues,
delay all the forms relating to the slots to the same position in
the build (with a new late-objdef source file) but execute all the
forms relating to specials, constants and primitive objects
themselves early.

This is probably the longest-standing SBCL issue, being logged in
the code by WHN 1999-01-31.  It's nice to fix it before we run out
of architectures to port to.