Remove TRACE-TABLE-OFFSET slot from CODE.
It stored the byte offset to the location where the code ended and
where the trace table started, but since there's no trace tables, it
was only used for getting the byte size of the code.
The code size is already stored in the CODE-SIZE slot, but in words
and rounded up to two words.
Store verbatim bytes instead and do the conversions when necessary.