Improve STRING= on x86[-64] by using memcmp() sometimes.
Also the unequal length pre-test, like in STRING-EQUAL, was missing.

This benchmarked twice as fast for strings of 20 characters in length
on two different pieces of hardware and approached 70x for huge strings.

Hopefully it will not be ruined with Unicode normalization, which
presumably occurs on creation of strings, not in every call to string=.