Improve the dumping of package tables during genesis.
- During cold-init before packages have been filled with symbols,
all symbols to be interned have their package slot set correctly.
This represents an inconsistent/impossible state normally,
but is immensely useful when examining a crash in early cold-init.

- Initial packages were intentionally oversized, but not all
of them gained additional symbols throughout the bootstrap process,
so this ended up being slightly wasteful, because resizing works
just fine, and we adjust them all again in 'save' anyway.

- The comment that "only symbols which actually have definitions ...
end up in the target packages" reflected an intent that did not
coincide with actual behavior. Extra symbols from package-data-list
affected the target. This is mostly fixed except for a hack that
assumes that every symbol in any package needed in "src/cold/warm"
is eventually needed. Notably, more than 200 symbols of SB-UNIX that
were possibly at some point manifest constants (e.g. READOWN) are
undefined and so do not appear now in the target.
Granted, package-data-list should be updated to reflect reality,
but the remainder of the problem is fixable with ease.

- *IN-PACKAGE-INIT* and *!DEFERRED-USE-PACKAGES* are eliminated.