Improve the OAOO-ness of DO-SYMBOLS, DO-{ALL,EXTERNAL}-SYMBOLS.
Iteration is somewhat slower, but now there are fewer places to touch
when fixing the bug related to simultaneous FIND-SYMBOL and rehash in
multithreaded code. Conforming code that uses these macros would not
encounter the rehash bug, due to constraints on traversal operations,
but a fix for the bug necessarily affects these macros.
So the trade-off seems fair considering the unlikelihood of package
iteration being a bottleneck in an application's performance.