Enhance the mechanism for recording full-calls in IR2.
This could be used to generate style-warnings for inline functions
and compiler-macros that are defined after their first use, in a way
that is potentially more robust than recording all functions for which
:WHERE-FROM become :ASSUMED after any file, and hooking the INLINE
proclamation to signal a warning.

The compile-time logic is sufficiently lightweight that there
is no reason to make it conditional on #!+sb-show,
and if disabled there is no cost at all anyway.