Workaround for lp# 1370836
Part of this change is sane and part is a total hack.
The reasonable aspect is that by separating 4 tests, we don't
see all of the :FUNCTION-LAMBDA-EXPRESSION test failing - only
the functions that failed the disassembly test.
The kludgy part is that this isn't even a deterministic fix
for the disassembler bug - it's just a randomization of the
memory addresses that happens to make it not crash now.