99 host compiler leaks, I got 'em
take one down, pass it around, there are probably still 99 of
them.  (There's probably a case for doing to GENSYM what we do
to BYTE and BOOLE, but this is good enough for now).

With this change, there are 8 files left in an x86-64 build with
differing outputs from sbcl and clisp as host compiler.

Several of them are, I think, to do with dumping of literal alien
type structures; one of them is a difference in the
ucdmisc binary database; one or two of them are even more
mysterious.