Rebind *HANDLER-CLUSTERS* in SIGNAL to prevent infinite recursion
Commit

  3f3407f0 "Fewer *HANDLER-CLUSTERS* special re-bindings in SIGNAL"

removed some of the re-bindings of *HANDLER-CLUSTERS* in SIGNAL as an
optimization. This lead to the current cluster not being removed
from *HANDLER-CLUSTERS* when testing condition types which in turn lead
to infinite recursive SIGNAL calls when a condition type test signaled
an error.

Re-bind *HANDLER-CLUSTERS* to not contain the current cluster before
applying condition type tests.

fixes lp#1378939