Change WRITE-TO-STRING from FLUSHABLE to UNSAFELY-FLUSHABLE.
An "unused" call of the function might have been intended to signal a
condition about unprintable objects, or perform random side-effects
in user-defined printers. But in practice, calls were not flushed,
because if the compiler-macro worked, the transform to STRINGIFY-OBJECT
was not flushable.

Incidentally, make STRINGIFY-OBJECT a known fun so that the compiled code
for the expansion of WRITE-TO-STRING avoids using RETURN-MULTIPLE.