Solve GC issues caused by interrupt handling.
Neither of the alloc_ functions in alloc.c are GC safe, which could be
fixed on x86oids, but is a bit more harder on precise GC platforms.
The most often used function, alloc_sap, can actually be removed,
because the sap can be stack allocated, since it's used for handling
interrupts.