more deterministic xc-host version of VALUES-SPECIFIER-TYPE
now that VALUES-SPECIFIER-TYPE is the only cached function which
accepts arbitrary lisp objects (and uses SXHASH), make an xc-host
version of it which simply uses an EQUAL hash table, and stores
everything during the XC build.  Strictly speaking this is not quite
correct, because of the theoretical possibility of multiple types
like (MEMBER #1=(FOO)) (MEMBER #2=(FOO)), but this is pretty unlikely
to happen.

I would have liked to write

  (macrolet ((values-specifier-type-body ...))
    #+sb-xc-host (let ((table (make-hash-table))) (defun ...))
    #-sb-xc-host (define-cached ...))

but that breaks during cross-compilation when one of the DEFINE-CACHED
components complains about not being able to be inlined on the grounds
that the lexical environment is too hairy.  Oh well.