Give perspicuous names to the object allocators used in Genesis.
The revised names ALLOCATE-HEADER+OBJECT and ALLOCATE-OBJECT
do exactly what they say, irrespective of whether an object
is thought of as boxed or unboxed.

COLD-FDEFINITION-OBJECT was working correctly only by dumb luck;
subtracting 1 when calling ALLOCATE-BOXED-OBJECT was wrong, because
the boxed allocator did not add a header, but round-to-even made it
accidentally come out right, and would not have if fdefns had an odd
physical length because 1- would have not been rounded up.
ALLOCATE-SYMBOL used the unboxed allocator which was just strange,
but correct in that it did add a header to 1- the specified length.