Better implementation of frame restarting on x86oids.
Instead of putting a specially constructed binding pair on the binding
stack save the binding stack pointer on the stack, which can be
accessed through the debug fun. This requires less code, doesn't need
unwinding and doesn't stop TCO, thus it's now enabled with the default
optimization qualities. With (and (> debug 0) (>= debug speed)) it
makes frames restartable, and with (and (= debug 3) (> debug speed))
it prevents TCO as well.

Incidentally, this fixes lp#1405456 caused by the sentinels.