Solve immediate offset issues on ARM.
When storing on stack and the offset is larger than 12 bits, adjust
CFP and then store it. Not the best way and will confuse the debugger,
but it is used in move VOPs which do not have temporary registers, and
most of the time this is not exercised.
When loading, load the offset into the destination register.

Fixes lp#1335303.