Fix INVALID-ARG-COUNT-ERROR on x86-64 with ud2-breakpoints (Darwin).
To my surprise, the "wtf?" branch of INTERNAL-ERROR was taken
because INTERNAL-ERROR-ARGS was made to return one hardcoded location
- dubiously named arg-count-sc instead of arg-count-sc-offset -
only when the trap kind was invalid-arg-count. But VERIFY-ARG-COUNT
didn't emit the new trap kind for UD2, nor use the older technique of
emitting RCX's sc-offset into the instruction stream after the break.