Show the right function in the backtrace from local tail calls.
The control may flow to a local tail call without any jumps, and if
just before the tail call there's a call to another function, the
return address saved on the stack will be of the tail function.
This confused the debugger and showed that the call to that function
was coming from the tail function, not from the main function.
Do inclusive PC comparison for non-escaped frames.