Introduce suitable "cleanup" code for conditional DX allocation.
  * When a control-flow branch which did not perform a conditional
DX allocation joins with one which did, we have to perform a
"dummy" allocation in order that de-allocation will work correctly
once we leave the DX environment.

  * Introduce %DUMMY-DX-ALLOC to go alongside %POP-VALUES and
%NIP-VALUES as stack-analysis-based "cleanup" functions.  This
function sets the saved stack pointer of a DX LVAR to the depth of
another stack-allocated LVAR, or to the top of the stack.

  * And arrange to have DISCARD-UNUSED-VALUES first do the discard
down to the intersection of the stack contents of the old and new
blocks, then dummy up any required DX LVARs.

  * This kills lp#1044465, at long last.