Reader enhancement from sbcl-devel (Nov. 2014) with better interface.
The primary use-case for this presently is to print line numbers
in log messages from user code, so there is one exported macro
named COMPILE-FILE-POSITION to do that. Future work includes
dumping a new artifact allowing annotation of backtraces by line#.

There is no switch to toggle collection of the data at read-time.
The slowdown in COMPILE-FILE is typically less than 3%, and
there is no impact on COMPILE (which can't report line numbers).