Allow COMPILE-FILE-POSITION in more situations
A way to understand the new logic is that supposing (hypothetically) that
COMPILE-FILE-POSITION were a function, not a macro, then previously it was
acceptable only where the call to it would occur at runtime and not sooner.
This patch makes it ok to use at read-time and/or compile-time.
Of course since it is a macro, the call always occurs in the compiler,
so the preceding is just a model for explaining the behavioral change.