Avoid type conflicts when transforming EXPT.
EXPT in some cases was expanded into an IF form one leg of which was
%MAKE-RATIO. Which caused type conflicts, and we usually don't
consider things like (or good-type bad-type) as conflicting, even
though it can catch a great deal of bugs. Hide %MAKE-RATIO behind
RECIPROCATE, this also reduces code size, since ratio creation isn't
really the case the transform is optimizing.