Use EXPAND-DS-BIND for most macro-like things.
- DEFTYPE, !DEF-TYPE-TRANSLATOR, DEFINE-ALIEN-TYPE-TRANSLATOR
- DEF-IR1-TRANSLATOR
- DEFINE-COMPILER-MACRO
- DEFINE-INSTRUCTION-MACRO
- DEFSETF and DEFCAS

Missing from the above list are destructuring-bind and "actual" macros,
on account of a few bootstrapping issues pending resolution.

This doesn't yet use the expander that is able to elide type tests.
It does however avoid all the (CDR (CDR (CDR ...))) nonsense.

A test of DEFTYPE in sb-introspect now expects a different result,
because we generally delete &WHOLE and &ENVIRONMENT variables
from lambda lists in their returned representation.