Fix STACK analysis in the presence of loops near DX allocation
  * STACK analysis for DX LVARs involves tracing the control flow
graph backwards from the "entry DX" node to each accessible USE of
the LVAR.  This is done via depth-first search, bounded by the
"lowest common dominator" of all of the blocks containing USEs of
the LVAR.

  * Leaving the pun aside, to prevent infinite loops due to the
depth-first search following around a loop in the CFG multiple
times, BACK-PROPAGATE-ONE-DX-LVAR would refuse to visit the same
block twice on its path from the environment start ("entry DX")
to any given USE.  This had the unfortunate side-effect of never
considering the LVAR to be live within the blocks that are part
of the loop unless they are also part of the straight-line path.

  * The correct thing to have done, of course, and what we do here
and now, is to track which paths between blocks have been taken,
not which blocks have been visited, and to refuse to take the same
path twice (that is, refuse to visit the same arc twice, not the
same node twice).

  * Originally reported by jasom in #lisp, entered into launchpad
by |3b|, narrowed down to a STACK analysis issue by stassats
in #sbcl, and caused by myself, as I wrote this part of the code
in the first place.  Mea culpa.