Clean up two instruction format definitions on x86-64.
The instruction format definitions for near jumps (both conditional and
unconditional) used to hack around a limitation of instruction sizes in
the disassembler to a maximum of 32 bits, using an ad hoc prefilter.
This limitation has been changed quite a while ago (in 0.9.6.16); now
the maximum is the machine word size. So finally simplify these
instruction format definitions.