Name anonymous lambdas without preserving useless junk.
In particular, DEFTRANSFORM would name its lambda as
 (LAMBDA (NODE &AUX (var1 (massive-hair)) (var2 (massive-hair)) ...))
which is an idiom that avoids parsing the body for declarations,
injecting a LET, injecting declarations pertinent to NODE in the outer
context, and others in the inner context, etc, etc.

Reduces 64-bit core size by about 360K, or half that on 32-bit.