ARG-COUNT-ERROR function serves no purpose - remove it.
This was potentially called by each use of a destructuring-bind,
and the theory was that it was cheaper than passing keyword arguments.
This is no longer true, and it's not worth having another random
error-signaling function just to convert positional to keyword args
in at most two places.
Incidentally, the declaration in 'fndb' without SB!KERNEL:: was bogus.