Compute and memoize the SYMBOL-HASH of all symbols loaded from fasls.
The motivation of this is to allow for a possible enhancement to CASE
statement macroexpansion. Symbol-hash can be used to divide-and-conquer,
wherein selection over a huge set of symbols becomes a decision tree:
  (if (logbitp 1 sym)
      (if (logbitp 2 sym)
          <more-cases>
Within a subtree of fewer than a small number of choices, the tests would
be linear ("if/else/else"). Bits can be picked by the macroexpander
to create a nicely balanced tree since hashes are stable across images.
e.g. if every symbol's 0th hash bit is 1, don't use that bit in the tests.

As a further optimization, we'd like for the CASE statement never to require
computing the hash. For compilation into memory, this is easy to ensure.
(While macroexpander side-effects are generally undesirable, this particular
effect would be invisible at the user level.) For loaded fasl files, the
loader can memoize the result, since doing so is transparent and never wrong.