Avoid forward references to PARSE-mumble-TYPE condition classes.
Also, since the debugger likes to print conditions,  ensure that
obsolete conditions are not as recursively error inducing.
The interpretation taken is that technically it is allowed not
to signal an error on encountering an obsolete condition instance
(for the purpose of printing), because undefined behavior occurred
as soon as the condition class was incompatibly redefined,
therefore any behavior (such as not signaling) is permissible.