Remove ineffective WITHOUT-INTERRUPTS form in code/timer.lisp
The form

  (sb!thread:make-thread (without-interrupts
                           (allow-with-interrupts
                             function))
                         ...)

was probably meant to call FUNCTION in the created thread with
interrupts disabled and local WITH-INTERRUPTS allowed, but had no effect
instead. The rationale probably was to mimic the behavior of
the (make-timer ... :thread {nil | THREAD}) cases which call the
function with interrupts disabled (as stassats points out).

However, since the current behavior has been used for a long time
without apparent problems, just remove the ineffective form instead of
changing it to exhibit the intended behavior. A slight change in the
MAKE-TIMER documentation string hopefully clarifies the description of
the (unchanged) behavior.