Temporary hack for newer Xcode until I think of the right thing.
If the shell environment contains SBCL_MACOSX_VERSION_MIN=10.6
then it overrides the autodetection. There are a few issues with
autodetection- that's it's obsolete, that Config is included before
it contains correct values, that we relied on silently-ignore-missing
include feature of Make, and the the default in the absence of a
known darwin9 platform is to assume lower-than-9. Too many problems.