Proclaim *DEFSTRUCT-HOOKS* special once only.
*DEFSTRUCT-HOOKS* has to be known before the first expansion of
!DEFSTRUCT-WITH-ALTERNATE-METACLASS, which happens in 'condition-boot'.
It would have worked to put the !DEFVAR there, but would be strange.
Another viable option would be a macro that proclaims a variable special
in the host and target, and makes it have a value before cold-init.
But since there isn't such a macro, just do an obvious workaround.