build: leave a *.o file that has all of the startup code.
https://bugs.launchpad.net/sbcl/+bug/1500628
The purpose of this is that you can build a program with C code and
other object files at link time.  Without this change you have to
dlopen the C code from a running SBCL program.  That doesn't work if
you need to link statically, only have a *.a or *.o file (not a *.so)
and is more fragile.

Patch by Fare/Tunes.