Put a compilation unit around make-host-1.
This is an attempt to make non-SBCL build hosts happy,
in addition to being the right thing to do anyway.
Problem was that the definitions of PARSE-{DEPRECATED,UNKNOWN}-TYPE
in cross-type are superseded by the ones in late-type,
which is apparently considered an error by some.
(The ones in condition-boot are also repeats, but we don't care)