Fix one of two bugs regarding handler-bind and interpreted code.
Re-ordering the pieces of the expansion of DEFUN is one fix.

The other is as yet unfixable: a HANDLER-BIND with (SATISFIES PRED)
as its condition test expands to a reference to #'PRED to signal a
style-warning if PRED is not defined at compile-time, causing the
interpreter to actually need the definition of PRED at expansion time.
I would hope nobody needs this to "work"