Remove unnecessary fype proclamation.
I think this stemmed from a time when people used these proclamations
to avoid "undefined function" warnings.
Now that there is a compilation-unit around make-host-1, this isn't
a great way to achieve that effect.  It would be fine to have a file
of only proclamations, but the ad-hoc approach is too inconsistent.