"Fix" MIPS Linux context register access
  * It turns out that Linux declares that MIPS context registers
are all 64 bits wide, even on 32-bit systems...  So, we deal with
that.  But then it turns out that 32-bit code on 64-bit systems
has all of its context register values sign extended, leading to
badness if we have to try and read a negative-fixnum from one of
the registers (which can happen with OBJECT-NOT-TYPE error trap
handling, for example).

  * "Fix", by masking down to 32 bits unconditionally.  We'll
almost certainly have to change this if we decide to support the
n32 calling convention (32-bit addresses, 64-bit registers), but
this should work for o32, which is what we use now.