Always output whitespacely-canonical headers from genesis
  * Clean up a minor irritation for those of us who don't always
take the time to do a full, clean build, especially for a change
that can be slammed, but do take the time to run
canonicalize-whitespace before a commit.  This keeps the headers
from changing due to a whitespace error, which means that they
will compare properly after one set has been canonicalized.