hppa: Fix funcallable-instance-tramp
  * The funcallable-instance-tramp is required to look and behave
like a SIMPLE-FUN object, for a couple of reasons.  One of them is
simply that a named call to a funcallable-instance occurs via the
closure-tramp, which expects that the closure-fun slot (the
funcallable-instance-trampoline slot) contains something that has
FUN-POINTER-LOWTAG and code starting at a specific offset from the
tagged address.  Even if the existing version happened to work (by
no means a sure thing), it is better to have it work by design
than by accident.